Add basic data/GPRS setup commands to modem class
This allows the manipulation of PDP contexts (setting APNs, etc) and can
initiate the first portion of a data session (the actual dialing). Not
much use as-is, but can be extended later for MMS sending functionality,
and with the use of pppd and a bit of hacking this can be used to
set up and control "normal" data network connections via the GSM modem.