Add workaround for broken modem +CMGR behaviour
Some ZTE modems sometimes issue incorrect +CMGR responses in PDU mode.
This attempts to work around the issues by added a default read status
to received SMS messages. This also includes a unittest containing a
semi-invalid SMS PDU; the encoder/decoder seems to handle it OK (except
for the sending number - might have to update the test and figure out a
way to detect when the PDU has this extra byte issue, and handle it).